Simplify the build process...
Ran into some difficulty getting `gulp-google-cdn` / `google-cdn` to convert my
`bower`-installed dependencies to CDN links, so no need for all this injecting
etc. I'll just build the dependencies into a bundle for now and CDN-ify them
later.

FWIW, `google-cdn` assumes that `dependencies` is a hash of package names and
`semver` definitions _only_. The package names must match _exactly_ the
name used in the `*-cdn-data` packages... unlike `normalize-css`. And in the
case of `components-modernizr`, the pacakge name is `modernizr` but the spec
is `components-modernizr#X.Y.Z` instead.